Create GATEembedded
This code is an example of GATE embedded in Java code. I'm using this code and some JAPE rules to detect and extract document structure. It still needs to be finished for this corpus but I think this code could be useful as an example for other uses and JAPE rules specific to the GATE application created by @penserbjorne 

The input is a folder with the input TXT files. The output is a folder with an XML file for each TXT file in the input folder. The annotations are indicated in the XML. The annotations can then be processed to be stored in a database or in a different representation.

If line 106 is uncommented, the GATE GUI is shown.

If line 113 is uncommented, a GATE application can be loaded.